Rename bedSearchParameters factory
I renamed this factory to `bedSearchApiParameters` to help clarify to use this factory to generate the necessary parameters needed by the API. I will create another factory specifically for bed search form which takes different attributes and values which the controller manipulates into something which the API will accept
