+++ e2e/tests/stepDefinitions/manage/bedspaceSearch.ts
@@ -3,7 +3,7 @@ import Page from '../../../../cypress_shared/pages/page'
 import DashboardPage from '../../../../cypress_shared/pages/temporary-accommodation/dashboardPage'
 import BedspaceSearchPage from '../../../../cypress_shared/pages/temporary-accommodation/manage/bedspaceSearch'
 import {
-  bedSearchParametersFactory,
+  bedSearchApiParametersFactory,
   bedSearchResultFactory,
   bedSearchResultsFactory,
 } from '../../../../server/testutils/factories'
@@ -20,7 +20,7 @@ Given('I search for a bedspace', () => {
   cy.then(function _() {
     const page = Page.verifyOnPage(BedspaceSearchPage)
 
-    const searchParameters = bedSearchParametersFactory.build({
+    const searchParameters = bedSearchApiParametersFactory.build({
       probationDeliveryUnits: [this.premises.probationDeliveryUnit.id],
       attributes: characteristicsToSearchAttributes(this.premises),
     })
