+++ server/controllers/temporary-accommodation/manage/bedspaceSearchController.test.ts
@@ -6,7 +6,7 @@ import { AssessmentsService } from '../../../services'
 import BedspaceSearchService from '../../../services/bedspaceSearchService'
 import {
   assessmentFactory,
-  bedSearchParametersFactory,
+  bedSearchApiParametersFactory,
   bedSearchResultsFactory,
   placeContextFactory,
   referenceDataFactory,
@@ -85,7 +85,7 @@ describe('BedspaceSearchController', () => {
       })
 
       it('renders with errors if the API returns an error', async () => {
-        const searchParameters = bedSearchParametersFactory.build()
+        const searchParameters = bedSearchApiParametersFactory.build()
         const placeContext = placeContextFactory.build()
 
         request.query = {
@@ -122,7 +122,7 @@ describe('BedspaceSearchController', () => {
 
     describe('when showing results', () => {
       it('renders the search results page with search results when given a search query', async () => {
-        const searchParameters = bedSearchParametersFactory.build()
+        const searchParameters = bedSearchApiParametersFactory.build()
 
         request.query = {
           ...searchParameters,
@@ -148,7 +148,7 @@ describe('BedspaceSearchController', () => {
       })
 
       it('updates the place context when given a search query', async () => {
-        const searchParameters = bedSearchParametersFactory.build()
+        const searchParameters = bedSearchApiParametersFactory.build()
         const placeContext = placeContextFactory.build()
 
         request.query = {
